// src/parser/types/GenericName.ts
enum GenericName {
  Bytes = 'bytes',
  Long = 'long',
  UnsignedLong = 'ulong',
  Float = 'float',
  Boolean = 'bool',
  Double = 'double',
  /**
   * shortcut to int32
   */
  Integer = 'int',
  Uint32 = 'uint32',
  Int32 = 'int32',
  Uint16 = 'uint16',
  Int16 = 'int16',
  Uint8 = 'uint8',
  Int8 = 'int8',
  NullTerminatedString = 'null_terminated_string',
  String = 'string'
}

export default GenericName;